created a new core.py file which contains all the core SeisFlows class definitions, which avoids some funky late type conversions in config defined a seisflows Base object which dictates the structure that all seisflows modules must have. This is intended to replace the individual Base objects defined in each of the sub-modules old Base modules (preprocess, postprocess) will now be named Default to better signify their role in the workflow (Base felt too abstract) removed Solver plugins. they weren't being used Removed error tools as they seemed overly abstract and were only used in one location. Replaced ParameterError with ValueError in the one instance of it Postprocess.default is now redefined with the new Base class definition, which no longer calls sys modules at the top of the script, removes the required property and defines required in __init__, inherits some properties from Base and explicitely defines super() where functions are being inherited.
